bpf: Protect each uintptr with runtime.KeepAlive
[ upstream commit 9f492a1b9ed8f8d665bc885fca02a7b9e521c96d ]

The Go's GC is unable to track objects referred by uintptr [1] [2].

A consequence of this is that the GC might collect such objects
when a blocking syscall is executed. This might corrupt memory of
other Go objects e.g. if the reclaimed memory is used for new objects and
the kernel modifies the objects in the context of the blocking syscall.

To prevent from this, we protect each pointer to objects with
runtime.KeepAlive() which marks the objects as alive at least until
runtime.KeepAlive() has been called.

[1]: https://github.com/golang/go/issues/13372#issuecomment-160655731
[2]: https://utcc.utoronto.ca/~cks/space/blog/programming/GoUintptrVsUnsafePointer

Co-developed-by: André Martins <andre@cilium.io>
Co-developed-by: Daniel Borkmann <daniel@iogearbox.net>

+++ pkg/bpf/map_linux.go
@@ -565,7 +565,7 @@ func (m *Map) DumpWithCallback(cb DumpCallback) error {
 		key:   uint64(uintptr(unsafe.Pointer(&key[0]))),
 		value: uint64(uintptr(unsafe.Pointer(&nextKey[0]))),
 	}
-	bpfCurrentKeyPtr := uintptr(unsafe.Pointer(&bpfCurrentKey))
+	bpfCurrentKeyPtr := unsafe.Pointer(&bpfCurrentKey)
 	bpfCurrentKeySize := unsafe.Sizeof(bpfCurrentKey)
 
 	bpfNextKey := bpfAttrMapOpElem{
@@ -574,7 +574,7 @@ func (m *Map) DumpWithCallback(cb DumpCallback) error {
 		value: uint64(uintptr(unsafe.Pointer(&value[0]))),
 	}
 
-	bpfNextKeyPtr := uintptr(unsafe.Pointer(&bpfNextKey))
+	bpfNextKeyPtr := unsafe.Pointer(&bpfNextKey)
 	bpfNextKeySize := unsafe.Sizeof(bpfNextKey)
 
 	for {
@@ -658,7 +658,7 @@ func (m *Map) DumpReliablyWithCallback(cb DumpCallback, stats *DumpStats) error 
 		key:   uint64(uintptr(unsafe.Pointer(&currentKey[0]))),
 		value: uint64(uintptr(unsafe.Pointer(&value[0]))),
 	}
-	bpfCurrentKeyPtr := uintptr(unsafe.Pointer(&bpfCurrentKey))
+	bpfCurrentKeyPtr := unsafe.Pointer(&bpfCurrentKey)
 	bpfCurrentKeySize := unsafe.Sizeof(bpfCurrentKey)
 
 	bpfNextKey := bpfAttrMapOpElem{
@@ -667,7 +667,7 @@ func (m *Map) DumpReliablyWithCallback(cb DumpCallback, stats *DumpStats) error 
 		value: uint64(uintptr(unsafe.Pointer(&nextKey[0]))),
 	}
 
-	bpfNextKeyPtr := uintptr(unsafe.Pointer(&bpfNextKey))
+	bpfNextKeyPtr := unsafe.Pointer(&bpfNextKey)
 	bpfNextKeySize := unsafe.Sizeof(bpfNextKey)
 
 	for stats.Lookup = 1; stats.Lookup <= stats.MaxEntries; stats.Lookup++ {
